+++ src/test/java/roomescape/controller/AdminControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.controller;
 
 
 import io.restassured.RestAssured;
